+++ lib/enver/binary_parser.ex
@@ -3,7 +3,6 @@ defmodule Enver.BinaryParser do
   `Enver.BinaryParser` is trying really hard to add value :P
   """
 
-  @type invalid :: Enver.invalid()
   @type opts :: Enver.parse_opts()
   @type val :: Enver.val()
   @type valid :: {:ok, String.t()}
@@ -12,12 +11,8 @@ defmodule Enver.BinaryParser do
   # API #
   #######
 
-  @spec parse(val(), opts()) :: valid() | invalid()
-  def parse("", _) do
-    {:error, "invalid binary"}
-  end
-
-  def parse(val, _) when is_binary(val) do
+  @spec parse(val(), opts()) :: valid()
+  def parse(val, %{type: :binary}) when is_binary(val) do
     {:ok, val}
   end
 end